Package and move modules to folder "code"
