--- D2_model_constraints.py
@@ -1,2 +1,0 @@
-def modelling_constraints():
-    return
