--- E3_indicator_calculation.py
@@ -1,2 +1,0 @@
-def evaluation():
-    return
