+++ code/D0_modelling_and_optimization.py
@@ -1,13 +1,12 @@
-import os.path
 import timeit
 import logging
 import oemof.solph as solph
 import oemof.outputlib as outputlib
-import pprint as pp
+
 try:
     from .D1_model_components import define, call_component, helpers
 except ImportError:
-    from D1_model_components import define, call_component, helpers
+    from code.D1_model_components import define, call_component, helpers
 
 class modelling:
     def run_oemof(dict_values):
