+++ code/D2_model_constraints.py
@@ -1,0 +1,2 @@
+def modelling_constraints():
+    return
