+++ code/E0_evaluation.py
@@ -4,8 +4,8 @@ try:
     from .E1_process_results import process_results
     from .E2_economics import economics
 except ImportError:
-    from E1_process_results import process_results
-    from E2_economics import economics
+    from code.E1_process_results import process_results
+    from code.E2_economics import economics
 
 class evaluation:
     def evaluate_dict(dict_values, results_main, results_meta, dict_model):
