+++ code/E3_indicator_calculation.py
@@ -1,0 +1,2 @@
+def evaluation():
+    return
