+++ code/E4_verification_of_constraints.py
