+++ code/F0_output.py
@@ -1,7 +1,7 @@
 try:
     from .F1_plotting import plots
 except ImportError:
-    from F1_plotting import plots
+    from code.F1_plotting import plots
 import pandas as pd
 import logging
